Refactor common methods to jwt transformer
